% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CarboA}
\alias{CarboA}
\title{Example dose response dataset}
\format{
A list containing two elements:
\describe{
  \item{x}{The drug concentration levels}
  \item{y}{The response values}
  ...
}
}
\usage{
CarboA
}
\description{
A simple dose response dataset used to illustrate the use of the
eadrm functions.
}
\keyword{datasets}
