% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea_funcs.R
\name{plot.eadrm}
\alias{plot.eadrm}
\title{Plot an eadrm object}
\usage{
\method{plot}{eadrm}(x, ...)
}
\arguments{
\item{x}{Object of class eadrm to plot.}

\item{...}{Additional arguments to plot. Currently ignored.}
}
\description{
Plots the data used to fit an eadrm object as well as the fitted
dose-response curve
}
\examples{
ea.fit <- eadrm(CarboA$y, CarboA$x)
plot(ea.fit)
}
\seealso{
\code{\link{eadrm}}
}
